fix jsch yes/no question which waits for response blocking output
This  fix https://sourceforge.net/p/jsch/mailman/message/26655308/
was already implemented this way but it was disabled in case there was no password.

I enabled it in either case (password or not). It seems to be working now.

This bug occurred on a remote cluster where I connected using a certificate without a password.